requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
yfinance>=0.2.40
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0